requirements.txt
mcp>=1.7.1
python-dotenv>=1.0.1
requests>=2.32.3
plotext>=5.2.8
rich>=13.9.2
click>=8.1.7
matplotlib>=3.7.0
